mamba/requirements.txt
astroid==2.3.3
autopep8==1.4.4
boto3==1.10.45
botocore==1.12.133
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
docutils==0.15.2
gitdb2==2.0.6
GitPython==3.0.5
google-auth==1.10.0
grpcio==1.26.0
grpcio-tools==1.26.0
idna==2.8
isort==4.3.21
jmespath==0.9.4
kubernetes==10.0.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
progressbar==2.5
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pyhelm==2.14.5
pylint==2.4.4
python-dateutil==2.8.0
python-dotenv==0.10.3
pyyaml>=4.2b1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.2.0
six==1.13.0
smmap2==2.0.5
supermutes==0.2.5
typed-ast==1.4.0
urllib3==1.24.2
virtualenv==16.7.9
websocket-client==0.57.0
wrapt==1.11.2
